refactor: 유효성 검증 Validator 관심사 분리
- carName 검증 로직과 tryNumber 검증 로직을 개별 클래스로 분리
- Validator 파일의 분리에 따라 테스트 코드 분리
