+++ src/validator/ValidateCar.js
@@ -1,20 +1,14 @@
-import Output from "./view/Output.js";
-import { ERROR_MESSAGE } from "./constants/message.js";
-import { CAR_SETTING, TRY_NUMBER_SETTING } from "./constants/setting.js";
+import Output from "../view/Output.js";
+import { CAR_SETTING } from "../constants/setting.js";
+import { ERROR_MESSAGE } from "../constants/message.js";
 
-export default class Validator {
+export default class ValidateCar {
   static carNamesAndCount(carNames) {
     this.#carNameLengthOverFive(carNames);
     this.#carNameLengthUnderOne(carNames);
     this.#carCountUnderTwo(carNames.length);
     this.#carCountOverHundred(carNames.length);
     this.#duplicateCarName(carNames);
-  }
-
-  static tryNumber(tryNumber) {
-    this.#tryNumberUnderOne(tryNumber);
-    this.#tryNumberOverHundred(tryNumber);
-    this.#tryNumberNotPositiveInteger(tryNumber);
   }
 
   static #carNameLengthOverFive = (carNames) => {
@@ -48,24 +42,6 @@ export default class Validator {
   static #duplicateCarName = (carNames) => {
     if (new Set(carNames).size !== carNames.length) {
       Output.error(ERROR_MESSAGE.duplicateCarName);
-    }
-  };
-
-  static #tryNumberUnderOne = (tryNumber) => {
-    if (tryNumber < TRY_NUMBER_SETTING.minTryNumber) {
-      Output.error(ERROR_MESSAGE.tryNumberUnderOne);
-    }
-  };
-
-  static #tryNumberOverHundred = (tryNumber) => {
-    if (tryNumber > TRY_NUMBER_SETTING.maxTryNumber) {
-      Output.error(ERROR_MESSAGE.tryNumberOverHundred);
-    }
-  };
-
-  static #tryNumberNotPositiveInteger = (tryNumber) => {
-    if (!Number.isInteger(tryNumber)) {
-      Output.error(ERROR_MESSAGE.tryNumberNotPositiveInteger);
     }
   };
 }
